Fix a few compiler warnings

+++ game/Doom/Base/i_main.cpp
@@ -928,8 +928,8 @@ void I_VramViewerDraw(const int32_t texPageNum) noexcept {
         // Figure out the y position and height for the texture cache entry.
         // Have to rescale in using a 24.8 fixed point format because the screen is 240 pixels high, and the texture cache is 256 pixels high.
         // I'm not sure why there is negative checks here however, texcoords are never negative?
-        int32_t y = ((int32_t) pTex->texPageCoordY << 8) * SCREEN_H / TCACHE_PAGE_SIZE;
-        int32_t h = ((int32_t) pTex->height        << 8) * SCREEN_H / TCACHE_PAGE_SIZE;
+        int16_t y = (int16_t)(((int32_t) pTex->texPageCoordY << 8) * SCREEN_H / TCACHE_PAGE_SIZE);
+        int16_t h = (int16_t)(((int32_t) pTex->height        << 8) * SCREEN_H / TCACHE_PAGE_SIZE);
 
         if (y < 0) { y += 255; }
         if (h < 0) { h += 255; }
@@ -937,8 +937,8 @@ void I_VramViewerDraw(const int32_t texPageNum) noexcept {
         y >>= 8;
         h >>= 8;
 
-        const int32_t x = pTex->texPageCoordX;
-        const int32_t w = pTex->width;
+        const int16_t x = pTex->texPageCoordX;
+        const int16_t w = pTex->width;
 
         // Draw all the box lines for the cache entry
         LIBGPU_setXY2(linePrim, x, y, x + w, y);                // Top
